Health display for mobs
Refactors HealthDisplay functionality to allows its use for either mobs or the player. Mobs' health displays follow them around.

Upon mob instantiation, the HealthDisplay's parent is set to that of the Canvas in the scene in order to render the health meters. This necessitates a reference to the HealthDisplay in EnemyController in order to destroy the HealthDisplay manually when the EnemyController is being destroyed.