Italic player prompt in log view
Asks the player how they want to dispatch the foe when the text input is activated.
This is done through a new private Class in `ThesisChatController` for representing italic text which isn't part of the narrative.

Also re-adds the word "kill" which was accidentally removed from the kill POST before
1cc813